backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pytest
mongomock
httpx
